% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_commit.R
\name{check_commit}
\alias{check_commit}
\title{check_commit}
\usage{
check_commit(
  root_directory = "/disks/PROJECT",
  group = "staff",
  commit = FALSE,
  push = FALSE
)
}
\arguments{
\item{root_directory}{A directory.}

\item{group}{A UNIX group.}

\item{commit}{A boolean.}

\item{push}{A boolean.}
}
\value{
A tibble
}
\description{
check_commit
}
