% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_black.R
\name{compute_brightness}
\alias{compute_brightness}
\title{compute_brightness}
\usage{
compute_brightness(colour)
}
\arguments{
\item{colour}{vector of any of the three kinds of R color specifications,
\emph{i.e.}, either a color name (as listed by colors()),
a hexadecimal string of the form "#rrggbb" or "#rrggbbaa" (see rgb).}
}
\description{
compute_brightness
}
\keyword{internal}
