% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ioslides.R
\name{create_ioslides}
\alias{create_ioslides}
\title{create_ioslides}
\usage{
create_ioslides(path, git = TRUE)
}
\arguments{
\item{path}{A path. If it exists, it is used.
If it does not exist, it is created, provided that the parent path exists.}

\item{git}{If TRUE, init git.}
}
\value{
Path to the newly created project or package, invisibly.
}
\description{
This function create an R project with ioslides
}
