% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rprofile.R
\name{mcprofile}
\alias{mcprofile}
\title{.Rprofile configuration as a function}
\usage{
mcprofile(
  given = NULL,
  family = NULL,
  email = NULL,
  orcid = NULL,
  .Renviron = "~/.Renviron",
  LANGUAGE = NULL,
  R_LIBS_USER = NULL,
  R_MAX_NUM_DLLS = NULL,
  TZ = NULL,
  GITHUB_PAT = NULL,
  GITLAB_PAT = NULL
)
}
\arguments{
\item{given}{A character string with the given names.}

\item{family}{A character string with the family name.}

\item{email}{A character string giving an e-mail address.}

\item{orcid}{A chracter string giving an ORCID.}

\item{.Renviron}{Path to the \code{.Renviron} file.}

\item{LANGUAGE}{Environment variable for language.}

\item{R_LIBS_USER}{Environment variable for user library.}

\item{R_MAX_NUM_DLLS}{Environment variable for the maximum number of DLLs to be loaded.}

\item{TZ}{Environment variable for time zone.}

\item{GITHUB_PAT}{Environment variable for GitHub access token.}

\item{GITLAB_PAT}{Environment variable for GitLab access token.}
}
\description{
.Rprofile configuration as a function
}
